requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
python-dotenv==1.0.0
openai>=1.0.0
pydantic-settings>=2.0.0
redis>=5.0.0
requests>=2.31.0
